New sprite atlases library implementation for multiple folder support
