+++ UI/Tabs/BuildingEditor/WTSBuildingLayoutEditor.cs
@@ -20,7 +20,8 @@ namespace Klyte.WriteTheSigns.UI
         private static WTSBuildingLayoutEditor m_instance;
         public static WTSBuildingLayoutEditor Instance
         {
-            get {
+            get
+            {
                 if (m_instance == null)
                 {
                     m_instance = FindObjectOfType<WTSBuildingLayoutEditor>();
@@ -32,7 +33,8 @@ namespace Klyte.WriteTheSigns.UI
         private string m_currentBuildingName;
         public string CurrentBuildingName
         {
-            get => m_currentBuildingName; private set {
+            get => m_currentBuildingName; private set
+            {
                 m_currentBuildingName = value;
                 ReloadBuilding();
             }
@@ -82,16 +84,16 @@ namespace Klyte.WriteTheSigns.UI
 
 
             AddLabel("", m_uiHelperHS, out m_labelSelectionDescription, out m_containerSelectionDescription);
-            KlyteMonoUtils.LimitWidthAndBox(m_labelSelectionDescription, (m_uiHelperHS.Self.width / 2), true);
+            KlyteMonoUtils.LimitWidthAndBox(m_labelSelectionDescription, (m_uiHelperHS.Self.width / 2), out UIPanel containerBoxDescription, true);
             m_labelSelectionDescription.prefix = Locale.Get("K45_WTS_CURRENTSELECTION") + ": ";
-            m_btnNew = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_New, OnCreateNewCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_NEWINCITY", false);
-            m_btnCopy = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Copy, OnCopyToCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_COPYTOCITY", false);
-            m_btnDelete = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Delete, OnDeleteFromCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_DELETEFROMCITY", false);
-            m_btnLoad = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Load, OnOpenGlobalFolder, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_OPENGLOBALSFOLDER", false);
-            m_btnExport = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Export, OnExportAsGlobal, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_EXPORTASGLOBAL", false);
-            m_btnSteam = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Steam, OnExportAsAsset, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_EXPORTTOASSETFOLDER", false);
-            m_btnReload = AddButtonInEditorRow(m_containerSelectionDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Reload, OnReloadDescriptors, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_RELOADDESCRIPTORS", false);
-            m_btnLock = AddButtonInEditorRow(m_containerSelectionDescription, CommonsSpriteNames.K45_Lock, OnLockSelection, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_LOCKCAMERASELECTION", false);
+            m_btnNew = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_New, OnCreateNewCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_NEWINCITY", false);
+            m_btnCopy = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Copy, OnCopyToCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_COPYTOCITY", false);
+            m_btnDelete = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Delete, OnDeleteFromCity, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_DELETEFROMCITY", false);
+            m_btnLoad = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Load, OnOpenGlobalFolder, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_OPENGLOBALSFOLDER", false);
+            m_btnExport = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Export, OnExportAsGlobal, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_EXPORTASGLOBAL", false);
+            m_btnSteam = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Steam, OnExportAsAsset, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_EXPORTTOASSETFOLDER", false);
+            m_btnReload = AddButtonInEditorRow(containerBoxDescription, Commons.UI.SpriteNames.CommonsSpriteNames.K45_Reload, OnReloadDescriptors, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_RELOADDESCRIPTORS", false);
+            m_btnLock = AddButtonInEditorRow(containerBoxDescription, CommonsSpriteNames.K45_Lock, OnLockSelection, "K45_WTS_BUILDINGEDITOR_BUTTONROWACTION_LOCKCAMERASELECTION", false);
             m_btnLock.color = LockSelection ? Color.red : Color.white;
             m_btnLock.focusedColor = LockSelection ? Color.red : Color.white;
             m_btnLock.pressedColor = LockSelection ? Color.red : Color.white;
@@ -140,26 +142,23 @@ namespace Klyte.WriteTheSigns.UI
         }
 
         private void OnOpenGlobalFolder() => ColossalFramework.Utils.OpenInFileBrowser(WTSController.DefaultBuildingsConfigurationFolder);
-        private void OnDeleteFromCity()
+        private void OnDeleteFromCity() => K45DialogControl.ShowModal(new K45DialogControl.BindProperties
         {
-            K45DialogControl.ShowModal(new K45DialogControl.BindProperties
+            message = Locale.Get("K45_WTS_PROMPTDELETEBUILDINGLAYOUT"),
+            showButton1 = true,
+            showButton2 = true,
+            textButton1 = Locale.Get("YES"),
+            textButton2 = Locale.Get("NO"),
+        }, (x) =>
+        {
+            if (x == 1)
             {
-                message = Locale.Get("K45_WTS_PROMPTDELETEBUILDINGLAYOUT"),
-                showButton1 = true,
-                showButton2 = true,
-                textButton1 = Locale.Get("YES"),
-                textButton2 = Locale.Get("NO"),
-            }, (x) =>
-            {
-                if (x == 1)
-                {
-                    WTSBuildingsData.Instance.CityDescriptors.Remove(m_currentBuildingName);
-                    ReloadBuilding();
-                }
+                WTSBuildingsData.Instance.CityDescriptors.Remove(m_currentBuildingName);
+                ReloadBuilding();
+            }
 
-                return true;
-            });
-        }
+            return true;
+        });
         private void OnCopyToCity()
         {
             WTSBuildingsData.Instance.CityDescriptors[m_currentBuildingName] = XmlUtils.DefaultXmlDeserialize<BuildingGroupDescriptorXml>(XmlUtils.DefaultXmlSerialize(CurrentEditingInstance));
